app/requirements.txt
asgiref==3.8.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
Django==4.2.16
django-admin-logs==1.0.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-sql-utils==0.6.1
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.4
email-validator==2.1.0.post1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql_server_has_gone_away==2.0.0
packaging==21.3
paramiko==2.12.0
passlib==1.7.4
password-strength==0.0.3.post2
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pysftp==0.2.9
python-decouple==3.8
python-dotenv==1.0.1
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==69.5.1
simplejwt==2.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
typing==3.7.4.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.12